Improve docs and move converter
